Updated deploy scripts
