+++ .gitignore
@@ -43,4 +43,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+**/.vscode
 ~*.xlsx
