+++ backend/helpers/deploy.ts
@@ -1,5 +1,10 @@
-import { runCommand } from './runCommand';
+import { execSync } from 'child_process';
 import { Command } from 'commander';
+
+const runCommand = (command: string, message: string = '') => {
+  console.log(`\n🤖`, message.length ? `${message}\n` : `Running command: ${command}\n`);
+  return execSync(command, { stdio: [process.stdin, process.stdout, process.stderr] });
+};
 
 const cdkDeploy = () => {
   try {
