+++ backend/lib/constants.ts
@@ -2,5 +2,4 @@ export const APP_NAME = 'schedular';
 
 export const DEFAULT_VALUES = {
   GITHUB_OWNER: 'eric-bach',
-  EMAIL: 'info@ericbach.dev',
 };
