+++ backend/package.json
@@ -7,16 +7,14 @@
   },
   "scripts": {
     "build": "tsc",
-    "build-all": "npm run cdk-clear-context && ts-node ./helpers/buildAll.ts",
-    "cinstall-all": "npm i && npm run cdk-clear-context && ts-node ./helpers/cinstallAll.ts",
-    "clean-all": "npm run cdk-clear-context && ts-node ./helpers/cleanAll.ts",
-    "install-all": "npm i && npm run cdk-clear-context && ts-node ./helpers/installAll.ts",
+    "build-all": "npm run cdk-clear-context && ts-node ./helpers/commands.ts \"npm run build\" \"node_modules, cdk.out, build\"",
     "cdk": "cdk",
     "cdk-clear-context": "npm run cdk -- context --clear",
+    "clean-all": "npm run cdk-clear-context && ts-node ./helpers/commands.ts \"rm -fR node_modules\" \"node_modules, cdk.out, build\"",
     "deploy": "npm run build && npm run build-all && ts-node ./helpers/deploy.ts",
-    "deploy-cicd": "npm run quick-deploy -- -s cicd",
     "deploy-backend": "npm run quick-deploy -- -s backend",
     "deploy-frontend": "npm run quick-deploy -- -s frontend",
+    "install-all": "npm i && npm run cdk-clear-context && ts-node ./helpers/commands.ts \"npm install\" \"node_modules, cdk.out, build\"",
     "quick-deploy": "ts-node ./helpers/deploy.ts",
     "seed": "npm run build && ts-node ./helpers/seed.ts",
     "test": "npm run build && jest --watchAll=false --coverage",
